refactor directory and files structure in app-analyses component
fix respective imports in app-reports
