+++ app-analyses/src/main/kotlin/com/usvision/analyses/analyzer/Analyzer.kt
@@ -1,0 +1,9 @@
+package com.usvision.analyses.analyzer
+
+import com.usvision.model.Visitable
+import com.usvision.model.Visitor
+
+interface Analyzer<out T> : Visitor {
+    fun getResults(): Map<Visitable, T>
+}
+
