+++ app-analyses/src/main/kotlin/com/usvision/analyses/analyzer/Count.kt
@@ -1,0 +1,7 @@
+package com.usvision.analyses.analyzer
+
+data class Count(
+    override val value: Any,
+    override val type: String,
+    override val unit: String
+): Measure
