+++ app-analyses/src/main/kotlin/com/usvision/analyses/analyzer/Measure.kt
@@ -1,0 +1,7 @@
+package com.usvision.analyses.analyzer
+
+interface Measure {
+    val value: Any
+    val type: String
+    val unit: String
+}
