+++ app-analyses/src/main/kotlin/com/usvision/analyses/analyzer/Measurer.kt
@@ -1,0 +1,3 @@
+package com.usvision.analyses.analyzer
+
+interface Measurer : Analyzer<Measure>
