+++ app-analyses/src/main/kotlin/com/usvision/analyses/analyzer/Relationship.kt
@@ -1,0 +1,5 @@
+package com.usvision.analyses.analyzer
+
+import com.usvision.model.Visitable
+
+data class Relationship(val with: Visitable)
