+++ app-analyses/src/main/kotlin/com/usvision/analyses/analyzer/RelationshipAnalyzer.kt
@@ -1,0 +1,3 @@
+package com.usvision.analyses.analyzer
+
+interface RelationshipAnalyzer : Analyzer<Relationship>
