+++ app-analyses/src/main/kotlin/com/usvision/analyses/analyzer/RelationshipsAnalyzer.kt
@@ -1,0 +1,3 @@
+package com.usvision.analyses.analyzer
+
+interface RelationshipsAnalyzer : Analyzer<Set<Relationship>>
