+++ app-analyses/src/main/kotlin/com/usvision/analyses/detector/ArchitectureInsight.kt
@@ -1,0 +1,3 @@
+package com.usvision.analyses.detector
+
+interface ArchitectureInsight
