+++ app-reports/src/main/kotlin/com/usvision/reports/planner/AnalyzerReusePlanner.kt
@@ -1,7 +1,7 @@
 package com.usvision.reports.planner
 
-import com.usvision.analyses.Analyzer
-import com.usvision.analyses.Detector
+import com.usvision.analyses.analyzer.Analyzer
+import com.usvision.analyses.detector.Detector
 import com.usvision.reports.utils.Plan
 import com.usvision.reports.utils.ReportRequest
 import kotlin.reflect.KClass
