+++ app-reports/src/main/kotlin/com/usvision/reports/utils/Plan.kt
@@ -1,7 +1,7 @@
 package com.usvision.reports.utils
 
-import com.usvision.analyses.Analyzer
-import com.usvision.analyses.Detector
+import com.usvision.analyses.analyzer.Analyzer
+import com.usvision.analyses.detector.Detector
 
 data class Plan(
     val analyzers: Set<Analyzer<Any>>,
