+++ app-reports/src/main/kotlin/com/usvision/reports/utils/ReportRequest.kt
@@ -1,6 +1,6 @@
 package com.usvision.reports.utils
 
-import com.usvision.analyses.Detector
+import com.usvision.analyses.detector.Detector
 import kotlin.reflect.KClass
 
 data class ReportRequest(
